Set power role to none while overheat mitigation is active
In the case of a short, CC detection should be disabled and VBUS voltage
should be set to 0V. The only way to guarantee this with a PD charger is
by setting the power role to none. This change also changes the expected
user behavior to disconnect USB for at least polling_freq * 2 seconds
after the port temperature clears the hysteresis threshold.

Bug: 67741167
Test: Simulate overheat conditions using fake_port_temp
Change-Id: I1f3c4095c559bf85d73821f7d8be4aa94f5e90df
Signed-off-by: Maggie White <maggiewhite@google.com>