usb: dwc3: Increase the TxFIFO resize factor
With the TxFIFO depth calculated with the current resize
factor, lot of flow control is observed with MTP file
transfer. As a result the file transfer speed is not
meeting the KPIs. Increasing this value so that
flow control is minimized and throughput is improved.

Change-Id: If64c2b6073df12edbcbb75fa08669f5268644e2a
Signed-off-by: Elson Roy Serrao <eserrao@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
(cherry picked from commit e837e98502133004fcf1ed2174e34a750ac8a1a1)
Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>
Signed-off-by: Tiktodz <ewprjkt@proton.me>
Signed-off-by: Kneba <abenkenary3@gmail.com>
Signed-off-by: dotkit <ewprjkt@proton.me>